feat: adds calculation of file and directory statistics

+++ src/commons/DirectoryNode.ts
@@ -23,6 +23,9 @@ export default class DirectoryNode implements IDirectoryTreeNode {
     private _sizeInBytes: number = 0;
     parent: DirectoryNode = null;
 
+    private _totalNumberOfDirectories: number = 0;
+    private _totalNumberOfFiles: number = 0;
+
     constructor(name: string, isComplete: boolean = false, parent?: DirectoryNode | null) {
         this.name = name;
         this.parent = parent || null;
@@ -39,10 +42,13 @@ export default class DirectoryNode implements IDirectoryTreeNode {
     addEmptyDirectory(name: string, isComplete: boolean = false): DirectoryNode {
         if (!this.directories.has(name)) {
             const newNode = new DirectoryNode(name, isComplete, this);
+            const prev = this.totalNumberOfDirectories;
+            this._totalNumberOfDirectories += 1;
             this.directories.set(name, newNode);
+            this.updateParentDirectoryNumber(prev);
         }
 
-        return  this.directories.get(name);
+        return this.directories.get(name);
     }
 
     setDirectory(name: string, node: DirectoryNode): DirectoryNode | null {
@@ -51,15 +57,22 @@ export default class DirectoryNode implements IDirectoryTreeNode {
             return null;
         }
         const prevSize = this.sizeInBytes;
+        const prevFiles = this._totalNumberOfFiles;
+        const prevDirs = this._totalNumberOfDirectories;
         if(this.directories.has(name)) {
             const childNode = this.getDirectory(name);
             this._sizeInBytes -= childNode.sizeInBytes;
         }
-        this._sizeInBytes += node.sizeInBytes;
         this.directories.set(name, node);
         node.parent = this;
 
+        this._sizeInBytes += node.sizeInBytes;
+        this._totalNumberOfDirectories = this.directories.size + node.totalNumberOfDirectories;
+        this._totalNumberOfFiles = this.files.size + node.totalNumberOfFiles;
+
         this.updateParentSize(prevSize);
+        this.updateParentFileNumber(prevFiles);
+        this.updateParentDirectoryNumber(prevDirs);
 
         return node;
     }
@@ -67,11 +80,14 @@ export default class DirectoryNode implements IDirectoryTreeNode {
     public addFile(name: string, data: FileData): void {
         const newNode = new FileNode(name, data);
         let prevSize = this._sizeInBytes;
+        let prevNumber = this.totalNumberOfFiles;
         this._sizeInBytes += data.size;
+        this._totalNumberOfFiles += 1;
 
         this.files.set(name, newNode);
 
         this.updateParentSize(prevSize);
+        this.updateParentFileNumber(prevNumber);
     }
 
     public removeFile(name: string): void {
@@ -81,6 +97,46 @@ export default class DirectoryNode implements IDirectoryTreeNode {
             this._sizeInBytes -= toRemove.data.size;
             this.files.delete(name);
             this.updateParentSize(prevSize);
+        }
+    }
+
+    private updateParentDirectoryNumber(prev: number) {
+        if(prev === this.totalNumberOfDirectories){
+            // no need to update parents
+            return;
+        }
+
+        let currentNode: DirectoryNode = this;
+        let currentParent: DirectoryNode = this.parent;
+
+        while(currentParent != null) {
+            let currentParentPrevSize = currentParent.totalNumberOfDirectories;
+            currentParent._totalNumberOfDirectories -= prev;
+            prev = currentParentPrevSize;
+            currentParent._totalNumberOfDirectories += currentNode.totalNumberOfDirectories;
+
+            currentNode = currentParent;
+            currentParent = currentParent.parent;
+        }
+    }
+
+    private updateParentFileNumber(prev: number) {
+        if(prev === this.totalNumberOfFiles){
+            // no need to update parents
+            return;
+        }
+
+        let currentNode: DirectoryNode = this;
+        let currentParent: DirectoryNode = this.parent;
+
+        while(currentParent != null) {
+            let currentParentPrevSize = currentParent.totalNumberOfFiles;
+            currentParent._totalNumberOfFiles -= prev;
+            prev = currentParentPrevSize;
+            currentParent._totalNumberOfFiles += currentNode.totalNumberOfFiles;
+
+            currentNode = currentParent;
+            currentParent = currentParent.parent;
         }
     }
 
@@ -104,12 +160,31 @@ export default class DirectoryNode implements IDirectoryTreeNode {
         }
     }
 
+    /**
+     * Removes directory from structure.
+     *
+     * @param name - directory name
+     * @return removed directory
+     */
     removeDirectory(name: string): DirectoryNode {
         const childNode = this.getDirectory(name);
-        this.directories.delete(name);
-        this._sizeInBytes -= childNode._sizeInBytes;
+        if(!childNode){
+            return null;
+        }
 
-        return this;
+        this.directories.delete(name);
+        const prevSize = this.sizeInBytes;
+        this._sizeInBytes -= childNode._sizeInBytes;
+        const prevNumOfDirs = this.totalNumberOfDirectories;
+        this._totalNumberOfDirectories -= (childNode.totalNumberOfDirectories + 1);
+        const prevNumOfFiles = this.totalNumberOfFiles;
+        this._totalNumberOfFiles -= childNode.totalNumberOfFiles;
+
+        this.updateParentDirectoryNumber(prevNumOfDirs);
+        this.updateParentFileNumber(prevNumOfFiles);
+        this.updateParentSize(prevSize);
+
+        return childNode;
     }
 
     hasChild(name: string): boolean {
@@ -134,5 +209,13 @@ export default class DirectoryNode implements IDirectoryTreeNode {
 
     public get sizeInBytes(): number {
         return this._sizeInBytes;
+    }
+
+    public get totalNumberOfFiles(): number {
+        return this._totalNumberOfFiles;
+    }
+
+    public get totalNumberOfDirectories(): number {
+        return this._totalNumberOfDirectories;
     }
 }
